relocations are done
