+++ containers/AuthFields/validation.js
@@ -1,28 +1,29 @@
 import { isStringEmpty, isEmail } from '../../libraries/validations';
 
 export default values => {
-  const errors = {};
-  const touched = true;
+  const errorsBuffer = {};
 
-  if (isStringEmpty(values.firstName)) {
-    errors.firstName = 'Required';
-  }
+  Object.entries(values).forEach(([key, value]) => {
+    if (isStringEmpty(value)) {
+      errorsBuffer[key] = 'Required';
+      return;
+    }
 
-  if (isStringEmpty(values.lastName)) {
-    errors.lastName = 'Required';
-  }
+    if (key === 'email' && !isEmail(value)) {
+      errorsBuffer[key] = 'Invalid email address';
+      return;
+    }
 
-  if (isStringEmpty(values.email)) {
-    errors.email = 'Required';
-  } else if (!isEmail(values.email)) {
-    errors.email = 'Invalid email address';
-  }
+    if (key === 'password' && values.password.length <= 3) {
+      errorsBuffer[key] = 'Must be at least 4 characters';
+    }
+  });
 
-  if (isStringEmpty(values.password)) {
-    errors.password = 'Required';
-  } else if (values.password.length <= 3) {
-    errors.password = 'Must be at least 4 characters';
-  }
+  // can be false or object
+  const errors = Object.keys(errorsBuffer).length > 0 ? errorsBuffer : false;
 
-  return { errors, touched };
+  return {
+    errors,
+    touched: true
+  };
 };
