+++ containers/Layout/index.js
@@ -1,7 +1,7 @@
 import { Helmet } from 'react-helmet';
 import PropTypes from 'prop-types';
-import App from '../containers/App';
-import Header from '../containers/Header';
+import App from './App';
+import Header from '../Header';
 
 const Default = props =>
   <App>
