+++ containers/PostList/index.js
@@ -1,13 +1,10 @@
-import { graphql } from 'react-apollo';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
 import { Link } from '../../routes';
 import PostUpvoter from '../PostUpvoter';
-import allPostsGql from './allPosts.gql';
+import { Section } from './styles';
+import connect from './data';
 
-const POSTS_PER_PAGE = 10;
-
-let PostList = ({
+const PostList = ({
   data: { allPosts, loading, _allPostsMeta },
   loadMorePosts,
   className
@@ -15,7 +12,7 @@ let PostList = ({
   if (allPosts && allPosts.length) {
     const areMorePosts = allPosts.length < _allPostsMeta.count;
     return (
-      <section className={className}>
+      <Section className={className}>
         <ul>
           {allPosts.map((post, index) =>
             <li key={post.id}>
@@ -44,7 +41,7 @@ let PostList = ({
               {loading ? 'Loading...' : 'Show More'}
             </button>
           : ''}
-      </section>
+      </Section>
     );
   }
   return <div>Loading</div>;
@@ -56,67 +53,4 @@ PostList.propTypes = {
   className: PropTypes.string.isRequired
 };
 
-PostList = styled(PostList)`
-  padding-bottom: 20px;
-
-  li {
-    display: block;
-    margin-bottom: 10px;
-  }
-  div {
-    align-items: center;
-    display: flex;
-  }
-  a {
-    font-size: 14px;
-    margin-right: 10px;
-    text-decoration: none;
-    padding-bottom: 0;
-    border: 0;
-  }
-  span {
-    font-size: 14px;
-    margin-right: 5px;
-  }
-  ul {
-    margin: 0;
-    padding: 0;
-  }
-  button:before {
-    align-self: center;
-    border-style: solid;
-    border-width: 6px 4px 0 4px;
-    border-color: #ffffff transparent transparent transparent;
-    content: "";
-    height: 0;
-    margin-right: 5px;
-    width: 0;
-  }
-`;
-
-export default graphql(allPostsGql, {
-  options: () => ({
-    variables: {
-      skip: 0,
-      first: POSTS_PER_PAGE
-    }
-  }),
-  props: ({ data }) => ({
-    data,
-    loadMorePosts: () =>
-      data.fetchMore({
-        variables: {
-          skip: data.allPosts.length
-        },
-        updateQuery: (previousResult, { fetchMoreResult }) => {
-          if (!fetchMoreResult) {
-            return previousResult;
-          }
-          return Object.assign({}, previousResult, {
-            // Append the new posts results to the old one
-            allPosts: [...previousResult.allPosts, ...fetchMoreResult.allPosts]
-          });
-        }
-      })
-  })
-})(PostList);
+export default connect(PostList);
