+++ containers/PostUpvoter/index.js
@@ -1,13 +1,12 @@
 import React from 'react';
-import { graphql } from 'react-apollo';
-import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import upvotePostGql from './upvotePost.gql';
+import connect from './data';
+import { UpvoteButton } from './styles';
 
-let PostUpvoter = ({ upvote, votes, id }) =>
-  <button onClick={() => upvote(id, votes + 1)}>
+const PostUpvoter = ({ upvote, votes, id }) =>
+  <UpvoteButton onClick={() => upvote(id, votes + 1)}>
     {votes}
-  </button>;
+  </UpvoteButton>;
 
 PostUpvoter.propTypes = {
   upvote: PropTypes.func.isRequired,
@@ -19,40 +18,4 @@ PostUpvoter.defaultProps = {
   votes: []
 };
 
-PostUpvoter = styled(PostUpvoter)`
-  background-color: transparent;
-  border: 1px solid #e4e4e4;
-  color: #000;
-
-  &:active {
-    background-color: transparent;
-  }
-
-  &:before {
-    align-self: center;
-    border-color: transparent transparent #000000 transparent;
-    border-style: solid;
-    border-width: 0 4px 6px 4px;
-    content: "";
-    height: 0;
-    margin-right: 5px;
-    width: 0;
-  }
-`;
-
-export default graphql(upvotePostGql, {
-  props: ({ ownProps, mutate }) => ({
-    upvote: (id, votes) =>
-      mutate({
-        variables: { id, votes },
-        optimisticResponse: {
-          __typename: 'Mutation',
-          updatePost: {
-            __typename: 'Post',
-            id: ownProps.id,
-            votes: ownProps.votes + 1
-          }
-        }
-      })
-  })
-})(PostUpvoter);
+export default connect(PostUpvoter);
