--- containers/SignInForm/apollo.js
@@ -1,15 +1,0 @@
-import { graphql } from 'react-apollo';
-import signInGql from './signinUser.gql';
-
-export const withMutation = graphql(signInGql, {
-  props: ({ mutate }) => ({
-    mutations: {
-      signInUser: ({ email, password }) =>
-        mutate({
-          variables: { email, password }
-        })
-    }
-  })
-});
-
-export default withMutation;
